Fixed a bug with possibility for starting two RPCs in parallel. (#87)
`notifyResultForCall`, `callOnSubscribeSuccess` and others call user code.
It's called through a `Handler`, so it's supposed to be async,
but the implementation is not guaranteed and it's possible
that the callback is executed immediately. If that happens,
that callback can trigger the next RPC call in the queue to start,
which means calling `startNextCall()` is incorrect any more.
This PR fixes it by always calling `startNextCallIfNotInProgress()` instead.